pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "brad-local-assistant"
version = "0.1.0"
description = "Brad - local-only AI meeting assistant"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
authors = [
  { name = "Brad contributors" }
]
dependencies = [
  "typer>=0.12.0",
  "rich>=13.7.1",
  "pydantic>=2.8.2",
  "pydantic-settings>=2.4.0",
  "sqlalchemy>=2.0.34",
  "gradio>=4.44.1",
  "faster-whisper>=1.1.0",
  "ctranslate2>=4.4.0"
]

[project.optional-dependencies]
vad = [
  "silero-vad>=5.1.2"
]
llm = [
  "llama-cpp-python>=0.2.90"
]
onnx = [
  "onnxruntime>=1.19.2",
  "optimum>=1.22.0"
]
dev = [
  "black>=24.8.0",
  "ruff>=0.6.4",
  "pytest>=8.3.2",
  "mypy>=1.11.1"
]

[project.scripts]
brad = "brad.cli:app"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
include = ["brad*"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "SIM", "RUF"]
ignore = ["E501"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
